Move error types from xerrors to relevant packages
The `internal/xerrors` package holds a couple of error types and nothing
else. These errors are each produced by only one package, and both `xerrors` and
the error producing packages are internal, so there's little reason for the
types to live in a separate package. The separation makes it less clear what the
errors are for, and `xerrors` feels a bit like the kind of "util" package we
generally discourage.

Move the error types from `internal/xerrors` into the packages that produce them
and remove the now empty `internal/xerrors`.
